Add default implementation of wrapper_fn
