+++ covalent/executor/__init__.py
@@ -32,7 +32,7 @@ import pkg_resources
 
 from .._shared_files import logger
 from .._shared_files.config import get_config, update_config
-from .base import BaseExecutor
+from .base import BaseExecutor, wrapper_fn
 
 app_log = logger.app_log
 log_stack_info = logger.log_stack_info
